03.StreamLit_Browser_ChatBot/requirements.txt
colorama==0.4.6
openai==2.17.0
python-dotenv==1.2.1
tiktoken==0.12.0
streamlit==1.54.0